Replace double with decimal in purchase method

+++ WeddingWebsite/Models/Registry/RegistryItemPurchaseMethod.cs
@@ -3,10 +3,10 @@
 public record RegistryItemPurchaseMethod(
     string Id,
     string Name,
-    double Cost,
+    decimal Cost,
     bool AllowBringOnDay = true,
     bool AllowDeliverToUs = true,
     string? Url = null,
     string? Instructions = null,
-    double DeliveryCost = 0
+    decimal DeliveryCost = 0
 );
